Reflect changes needed for snap variant
When Nextcloud has been installed via Snap, it needs other parameters than it pulls from the config file due to the usage of sockets. The config file reports the sockets under /tmp/sockets/, but the real host directory is /tmp/snap.nextcloud/tmp/sockets.

This check should ultimately be added to notify_push:setup (which should also go as far as installing the unit file and starting the service so that it's a more seamless process for the user).