feat(Dockerfile): switch cmd to entrypoint
By using `ENTRYPOINT` for the default command, it is possible to extend the command with arguments by using `CMD` in a custom Dockerfile. As the command itself doesn't need to be changed/overwritten with this approach, changes to the location of the binary or command will not break these configurations in the future.

Signed-off-by: Joda Stößer <git@simjo.st>